Emby: bump to 4.2.1.0 and fix logging

+++ spk/emby/src/installer.sh
@@ -117,8 +117,11 @@ preupgrade ()
 
 postupgrade ()
 {
-	# Restore some stuff
-	rm -rf ${INSTALL_DIR}/var
-	mv ${TMP_DIR}/${PACKAGE}/var ${INSTALL_DIR}/
+	# Restore user data
+	BACKUP=${TMP_DIR}/${PACKAGE}/var
+	if [ -d $BACKUP ]; then
+		rm -rf ${INSTALL_DIR}/var
+		mv ${TMP_DIR}/${PACKAGE}/var ${INSTALL_DIR}/
+	fi
 }
